+++ src/SendIt/src/Config/MailerConfig.php
@@ -36,14 +36,6 @@ final class MailerConfig extends InjectableConfig
         return $this->config['from'];
     }
 
-    /**
-     * @deprecated since v2.9.
-     */
-    public function getQueuePipeline(): ?string
-    {
-        return $this->getQueue();
-    }
-
     public function getQueue(): ?string
     {
         return $this->config['queue'] ?? $this->config['pipeline'] ?? null;