+++ tests/app/src/Bootloader/AppBootloader.php
@@ -18,7 +18,6 @@ use Spiral\App\Controller\AuthController;
 use Spiral\App\Controller\InterceptedController;
 use Spiral\App\Controller\TestController;
 use Spiral\App\Interceptor;
-use Spiral\App\User\UserRepository;
 use Spiral\App\ViewEngine\TestEngine;
 use Spiral\Bootloader\DomainBootloader;
 use Spiral\Bootloader\Http\JsonPayloadsBootloader;
@@ -27,7 +26,6 @@ use Spiral\Bootloader\Views\ViewsBootloader;
 use Spiral\Core\Core;
 use Spiral\Core\CoreInterface;
 use Spiral\Core\InterceptableCore;
-use Spiral\Domain\CycleInterceptor;
 use Spiral\Domain\FilterInterceptor;
 use Spiral\Domain\GuardInterceptor;
 use Spiral\Domain\PipelineInterceptor;
@@ -44,7 +42,6 @@ class AppBootloader extends DomainBootloader
     ];
 
     protected const INTERCEPTORS = [
-        CycleInterceptor::class,
         GuardInterceptor::class,
         FilterInterceptor::class
     ];
@@ -61,7 +58,6 @@ class AppBootloader extends DomainBootloader
     }
 
     public function boot(
-        \Spiral\Bootloader\Auth\AuthBootloader $authBootloader,
         RouterInterface $router,
         PermissionsInterface $rbac,
         ViewsBootloader $views,
@@ -69,8 +65,6 @@ class AppBootloader extends DomainBootloader
         JsonPayloadsBootloader $json,
         PipelineInterceptor $pipelineInterceptor
     ): void {
-        $authBootloader->addActorProvider(UserRepository::class);
-
         $rbac->addRole('user');
         $rbac->associate('user', '*');